user_profile_modal: Move title and tab switcher to a fixed header.
While going through the full profile of a user, we expect to find
that specific user's in-depth details. Thus, we should have a fixed
title displaying the name at the top of the modal to establish the
connection of the profile data with the user.

We also fix the tab switcher between "Profile/Streams/User groups" to
the top of the modal, for the added convinience of switching across
the three options at any point of time.

We also remove the no-fields class and it's corresponding CSS, since
there is no <hr> separator now for separating the custom profile
fields and neither do we reduce the height of the full profile modal
if no custom fields are present.